refactor(advanced): 쿠폰 관리 기능의 훅스 통합 및 경로 조정
- useCouponService, useCouponForm, useSelectedCoupon 훅스를 features/coupon-management로 이동
- CouponSelector 및 CouponTab 컴포넌트에서 관련 훅스 사용 방식 개선
