+++ src/advanced/features/coupon-management/index.ts
@@ -1,0 +1,3 @@
+export { useCouponService } from './model/useCouponService';
+export { useCouponForm } from './ui/useCouponForm';
+export { useSelectedCoupon } from './ui/useSelectedCoupon';
